Monster Experience: Continued
Ooops ... I need to adjust the xp in the computed score for ML/(PL+2).
In the end game, the xp you enter in r_info nearly gets doubled. So the
original xp values don't look so bad now, but quite a few could use a
boost, especially on the really hard uniques.