Wizard: Command to view shop stock
My CL26 HM just bought his 3rd spellbook from the BM, and I was curious
whether this was luck or a problem. So I made the wiz_obj statistics
report generally available. This simply creates 1000 objects using a
supplied creation function, adds them to an inventory, sorts and
reports. Looks like 1 in 3000 or so to get a spellbook at CL26. And to
get *my* spellbook? I guess 1 in 50000 with all the spell realms. So
quite lucky :)