oook report: vortex elemental explosion messed up
Due to nested calls to format(), a common and hard to detect error.
format() uses static global memory so you, the programmer, must be
completely finished with the results of one call to format() before
issuing another. Of course, this can be hard to detect when you call a
nested function ...

Other issues: The powers UI is very old and needs a rewrite. It cannot
handle color formatting codes.