oook report: block switching to wild_mode
Geez, Hugo must have reported this umpty-million times ;)

Well, if a monster is more distant than MAX_SIGHT, you can escape.
If not, then if the monster has LOS on you, you can no longer escape.
If no LOS, then you can escape if the distance to the monster exceeds
its "aaf". For metallic blue centipedes, this is a mere 6. MAX_SIGHT
is 20. So if you can block LOS to some non-observant monster, then you
can escape, and that is not a bug.

However, it was indeed a bug if you could escape while an awake monster
had you in sight. This should block wild_mode no matter what the aaf.

Sorry Hugo, I got this eventually (I think).