Paralysis: Ease up early game percentages
This, at least, will give careless players a slight chance of
recovering. Some monsters like Morganna really do need to be toned down,
being already incredibly nasty. The purpose of the new FA system is not
to squish noobs ... There's no sport in that. Instead, it is to change
paralysis from its old all-or-nothing design of insta-death or nil
effect. Paralysis means the player loses 1d3 moves (moves!, not timer
ticks which screw end game play). With FA, this is 1 move in general.
The only thing players need manage is chain-paralysis where on the very
turn you miss, the monster attacks, hits, triggers paralysis, and the
player fails their save. If this is happending 30 times in a row, then
the player is doing something wrong, or the monster design is messed up
(as with, say, a +10 speed monster with 4 100% paralysis blows).