Sanity Stick: wild_magic should *not* summon cyberdemons
On the one hand, I've played a long time never knowing this was even
possible. On the other, it's the exact kind of design that I hate:
screwing the player for something they have to do. A mage *has* to cast
spells. Killing them for a random failure (no matter how improbable), is
just dumb.

Bad effects for spell failure are fine. The low level chaos failures
are OK (teleport player, create traps, ... even disenchantment is ok).
But cyberdemons? C'mon :)