project: effects should not project_p more than once
This was recently reported on oook, and I've also noticed weirdness with
BR_STORM for a long time. Recently, Kavlax did a 500 nexus breath on me.
I really should have looked into this sooner ... apologies.

This bug (feature?) seems to have always existed. project() basically
iterates all "grids" in the projection. Just a loop. For each grid, it
sees if this grid contains the player, and then calls project_p(). But
it keeps on looping. So, if project_p *moves the player* to an as yet
to be processed projection grid, then the player will indeed get double or
even triple the fun.

I would call this a bug. Especially with BR_NEXUS(250) and BR_STORM(300).
It might be OK with a weak BR_GRAVITY. I have no idea if it is intentional
or not. Probably it is. But it is too harsh a mechanic and too non-predictable
to stay in poscheng.

Hard to test my fix though ... seems like it should work.