See Invisible: Add design spreadsheet
This will enable interested players to evaluate the design of this
mechanic. A build with poor searching skills will indeed be
disadvantaged. Mighty vs Nimble is 7 skill points. Human vs Hobbit is 12
skill points. And class is most important of all: Warriors are 14,
Rogues are 32 and Archaeologists are a whopping 50. In monster mode, try
a beholder.

The impression may be misleading, but one report of 1/20 odds really
should have had 20% odds with a single SI. Add Elven cloak +4 and
another SI and odds go to 75%. You can start with poor searching skills
and overcome that with magic. And of course, Telepathy (usually) helps
by at least providing fuzzy awareness.

I think the design is working as intended. I didn't test, but
-DetectMonsters should always reveal invisible foes without a skill
check. Players can use this *and* stealth to deal with them as well.