Monster Lore: When to display Damage
I made a spreadsheet to attempt to understand the original design. I'm
still not sure I do ... Perhaps melee has gone up alot since it was
written, but generally, you simply needed 100 for everything, and then
only because we enforced limits. I tweaked slightly for now, and will
come back to this later.