remove clouded monsters
Despite all my hard work to balance these contributions, I'm removing
them out of personal disgust with the contributor.

This might cause upgrade problems from 7.0.1:
[1] Uncompleted quests for removed monsters will become Cancelled once
the player reaches the quest level.
[2] Completed quests for removed monsters are still completed. They will
print as a quest to "Kill Removed Monster" on the character sheet.
[3] If a wanted monster is removed before defeated, players will not be
able to get that prize. If after defeated, the removed monster simply
won't show up in the knowledge lists.
[4] If the player is on a level with a removed monster and then
upgrades, the monster will be deleted from the level. Hopefully it was
not an active QUESTOR.