revert haunted house
Original was better. I'm pretty sure the proposed change was aiming at a
trivial quest doable immediately after the Wargs (purchase !DetectInvis)
for the sole purpose of getting a =Combat 10 minutes into the game. This
enables power starts (sometimes) making it easier to skip the early
game.

I've tried =Combat on early quests (Wargs) and decided it was a mistake.
I don't agree with skipping the early game. I also don't agree with
trivializing quests. Topi's version is indeed a bit over the top, but at
least it is challenging. I changed the reward to _BanishEvil, a L60
effect. Some players will do quests just for the challenge, not the
reward. But if we need to entice, then I'm open to improved reward
suggestions. Perhaps a rand-art?