restore Killing Fields to Telmora quest sequence
The Mountain Shrine was a good quest, imo. I'm removing it because I am
a bastard ;) Anyway, The Killing Fields is also a good quest. I don't
think moderate length quests are problematic. Nobody complains about The
Cloning Pits or The Old Castle, both of which are considerably longer.
And a !Life is worth getting ... You might only find 3 of these all game
long.