new quest for angwil
Replace Sand Pits -> Dragon Quest with The Renegade Sorcerer. Here's a
short quest for those that like that sort of thing. L45 devices and a
high spellbook await the brave, and sorcerers are a bit fragile. But
don't expect melee to always work unless you are particularly stealthy!
Rewards a =Speed usually at +3, but occasionally as high as +5.

Sand Pits was OK, but actually rather easy. Dragon Quest was a
monstrosity of object spam. Please, nothing more like that!