rooms: fix bug with default letters in v_info
Namely, FORMATION rooms allow 0 to 9 to refer to the 10 chosen formation
races (ordered by power descending). Usually, one only explicitly
defines the 0 letter to set the formation type. The problem is that 8
and 9 are very strong vault squares wrt object generation. Since they
are typically absent in the room template, looking up the default
character is quite the spectacular bug!

Paranoia wrt to @ and AMBUSH encounters.