ironman_rooms: remove
With the new vault changes, this option will no longer work. Perhaps we
could add something like a vault_pct birth option. This option would
simply scale the vault probabilities by the indicated percentage (from 0
to 100 to reduce the frequency of vaults, and from 100 to 5000 to
increase the frequency of vaults.) Just an idea ...