Anti-magic Caves: Better spell ai
Rather than super-buffing certain monsters, let the spell ai choose a
spell using the normal allocation. If a magical spell is chosen, then it
is rejected and the monster moves or attacks instead. For example,
Ghatanathoa spells 15% and breathes 23.5% normally. Inside the
Anti-magic caves, it would breathe 100% the old way, which was quite
deadly! The Anti-melee caves has no such hidden bias.