Remove quickmode birth option
This was an experimental option I added, but I'm not happy with its
effects. The goal was to make a shorter game, but, in reality, it only
made the early game shorter and more skippable. Since one of my main
goals for the last 6 years has been to rescue the early game, making it
more interesting, more playable, more balanced and more enjoyable, I
really feel the need to revert this option. Call it a failed experiment
with unintended consequences.