Performance: Speed up _bounty_uniques in _birth_finalize
This was slowed down by efforts to reduce the number of uniques per
level during level generation. In this case, summon_specific_who is set
to SUMMON_WHO_NOBODY. Need a new value for SUMMON_WHO_BIRTHER that
avoids the attempt to limit uniques. After all, we want 20 bounty
uniques while the code tries to limit uniques to just 4 or 5 per level!