Revert no_selling
This was added as a cherry pick from one of the FooPosChengband clones.
However, this option makes very little sense to me, and I'd rather not
clutter up the code, the documentation and the user interface with many
poor quality options. Players are free to sell or to not sell without
any code changes at all ... I suppose players might still want the lore
benefits from selling items without the gold benefits? However, I think
many players will expect to be able to turn this option on and have the
gold amounts actually increase so that they effectively gain a larger
pack size. Since there is absolutely no effect on gold drops, having
this option is misleading. I've seen at least one player walking around
with every single piece of armor rusted to the max, with a consequent
extremely low ac, and without the wherewithal to repair this fact at the
armory. And he shortly died to paralysis and considered that to be a
bug?! (I'm guessing as to the cause, since no dump or explanation was
actually given. Abnormally low AC is probably the only way to die from
paralysis if you have FA.)