Device Lore: Learn Power and Difficulty thru Use
For devices with direct damage, the player can now learn the device
power and damage numbers thru direct use. If the player inflicts the
entire damage amount on at least one monster, then lore is acquired.

For healing devices, the player learns if they actually need the entire
amount healed. Careful players could always do this since the game shows
the player's hp both before and after using the device (modulo
regeneration, of course).

For all devices, failing to use the device properly will occasionally
reveal the device's difficulty and fail rate.

Should the player learn both power and difficulty thru use, it is as if
the device is *identified* and they should see mana and casting cost as
well.